+++ src/usecases/retrieveOrganizations.js
@@ -1,32 +1,5 @@
-import logger from "../../logger";
-
-const retrieveOrganizations = ({ getDb }) => async () => {
-  const db = await getDb();
-  try {
-    const organizations = await db.any(
-      `SELECT
-        id as id,
-        name as name,
-        status as status
-      FROM
-        organization`
-    );
-
-    return {
-      organizations: organizations.map((organization) => ({
-        id: organization.id,
-        name: organization.name,
-        status: organization.status,
-      })),
-      error: null,
-    };
-  } catch (error) {
-    logger.error(error);
-    return {
-      orgznizations: null,
-      error: error.toString(),
-    };
-  }
+const retrieveOrganizations = ({ getRetrieveOrganizations }) => async () => {
+  return await getRetrieveOrganizations()();
 };
 
 export default retrieveOrganizations;
